added database and made tradeclientimpl with offsets (#8)

+++ src/main/scala/org/github/insider/polymarket/domain/Trade.scala
@@ -6,8 +6,8 @@ case class Trade(
   wallet: String,
   token: Token,
   side: Side,
-  size: Double,
-  price: Double,
+  size: BigDecimal,
+  price: BigDecimal,
   timestamp: Long
 )
 
@@ -19,8 +19,8 @@ object Trade {
         tokenId   <- c.downField("asset").as[String]
         outcome   <- c.downField("outcome").as[Outcome]
         timestamp <- c.downField("timestamp").as[Long]
-        size      <- c.downField("size").as[Double]
-        price     <- c.downField("price").as[Double]
+        size      <- c.downField("size").as[BigDecimal]
+        price     <- c.downField("price").as[BigDecimal]
         side      <- c.downField("side").as[Side]
       } yield Trade(wallet, Token(outcome, tokenId), side, size, price, timestamp)
     }
